backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.13
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.1
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
email-validator==2.2.0
